iced_winit: Split interface Application with implementation Platform. iced: Make user directly implement Application trait to remove wrapper boilerplate (Downside: Needs user to explicitly mixin associated type Renderer = platform::Renderer).
