+++ winit/src/lib.rs
@@ -2,10 +2,11 @@ pub use iced_native::*;
 pub use winit;
 
 pub mod conversion;
-
+mod platform;
 mod application;
 
 pub use application::Application;
+pub use platform::Platform;
 
 // We disable debug capabilities on release builds unless the `debug` feature
 // is explicitly enabled.
